+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+## Added
+- Error messages for private mode authority validations
+- `InstructionData` constant for `changePrivateModeAuthority` instruction
 
 ## [v2.10] - 2025-11-24